chore: rename `eth-price` package
