+++ packages/price-feed/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@blobscan/eth-price",
+  "name": "@blobscan/price-feed",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "scripts": {
